+++ actionpack/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Rails 3.2.0 (unreleased) ##
 
+*   Log "Filter chain halted as CALLBACKNAME rendered or redirected" every time a before callback halts. *José Valim*
+
 *   You can provide a namespace for your form to ensure uniqueness of id attributes on form elements.
     The namespace attribute will be prefixed with underscore on the generate HTML id. *Vasiliy Ermolovich*